requirements.txt
Django==1.9
Markdown==2.6.5
appnope==0.1.0
coverage==4.0.3
decorator==4.0.6
django-filter==0.11.0
djangorestframework==3.3.2
gnureadline==6.3.3
ipython==4.0.1
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
redis==2.10.5
simplegeneric==0.8.1
traitlets==4.0.0
wsgiref==0.1.2
django-cors-headers==1.1.0
